# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LLM Tool Bridge API",
    "description": "An OpenAI-compatible proxy that adds tool/function calling capabilities to any LLM. This proxy intelligently manages multi-turn conversations, handles tool execution, and provides seamless integration with existing OpenAI SDK clients.",
    "version": "1.0.0",
    "contact": {
      "name": "LLM Tool Bridge Support"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3500",
      "description": "Main Proxy Server"
    },
    {
      "url": "http://localhost:3501",
      "description": "Admin API Server"
    }
  ],
...
  "paths": {
    "/v1/chat/completions": {
      "post": {
        "tags": ["Chat"],
        "summary": "Create chat completion",
        "description": "Creates a chat completion with support for tool/function calling. The proxy automatically handles tool execution across multiple rounds, managing the conversation flow between the LLM and tool responses. Compatible with OpenAI SDK.",
        "operationId": "createChatCompletion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              },
              "examples": {
                "simple": {
                  "summary": "Simple chat without tools",
                  "value": {
                    "model": "gpt-3.5-turbo",
                    "messages": [
                      {"role": "user", "content": "Hello, how are you?"}
                    ]
                  }
                },
                "withTools": {
                  "summary": "Chat with tool calling",
                  "value": {
                    "model": "gpt-3.5-turbo",
                    "messages": [
                      {"role": "user", "content": "What's the weather in San Francisco?"}
                    ],
                    "tools": [
                      {
                        "type": "function",
                        "function": {
                          "name": "get_weather",
                          "description": "Get the current weather",
                          "parameters": {
                            "type": "object",
                            "properties": {
                              "location": {"type": "string"}
                            },
                            "required": ["location"]
                          }
                        }
                      }
                    ]
                  }
                },
                "streaming": {
                  "summary": "Streaming response",
                  "value": {
                    "model": "gpt-3.5-turbo",
                    "messages": [
                      {"role": "user", "content": "Tell me a story"}
                    ],
                    "stream": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionStreamResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable - No LLM target configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {"bearerAuth": []},
          {}
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["model", "messages"],
        "properties": {
          "model": {
            "type": "string",
            "description": "ID of the model to use"
          },
          "messages": {
            "type": "array",
            "description": "List of messages comprising the conversation",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            }
          },
          "tools": {
            "type": "array",
            "description": "List of tools/functions available for the model to call",
            "items": {
              "$ref": "#/components/schemas/Tool"
            }
          },
          "tool_choice": {
            "oneOf": [
              {"type": "string"},
              {"type": "object"}
            ],
            "description": "Controls tool selection: 'auto', 'none', or specific tool"
          },
          "temperature": {
            "type": "number",
            "minimum": 0,
            "maximum": 2,
            "description": "Sampling temperature (0-2)"
          },
          "max_tokens": {
            "type": "integer",
            "description": "Maximum tokens to generate"
          },
          "stream": {
            "type": "boolean",
            "description": "Whether to stream the response"
          },
          "top_p": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "Nucleus sampling parameter"
          },
          "frequency_penalty": {
            "type": "number",
            "minimum": -2,
            "maximum": 2,
            "description": "Frequency penalty"
          },
          "presence_penalty": {
            "type": "number",
            "minimum": -2,
            "maximum": 2,
            "description": "Presence penalty"
          }
        }
      },
      "ChatMessage": {
        "type": "object",
        "required": ["role", "content"],
        "properties": {
          "role": {
            "type": "string",
            "enum": ["system", "user", "assistant", "tool"],
            "description": "Message role"
          },
          "content": {
            "type": "string",
            "description": "Message content"
          },
          "name": {
            "type": "string",
            "description": "Name of the function (for tool messages)"
          },
          "tool_calls": {
            "type": "array",
            "description": "Tool calls made by the assistant",
            "items": {
              "$ref": "#/components/schemas/ToolCall"
            }
          },
          "tool_call_id": {
            "type": "string",
            "description": "ID of the tool call this message responds to"
          }
        }
      },
      "Tool": {
        "type": "object",
        "required": ["type", "function"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["function"],
            "description": "Tool type (currently only 'function' is supported)"
          },
          "function": {
            "$ref": "#/components/schemas/ToolFunction"
          }
        }
      },
      "ToolFunction": {
        "type": "object",
        "required": ["name", "description"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Function name"
          },
          "description": {
            "type": "string",
            "description": "Function description for the model"
          },
          "parameters": {
            "type": "object",
            "description": "JSON Schema for function parameters"
          }
        }
      },
      "ToolCall": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for this tool call"
          },
          "type": {
            "type": "string",
            "enum": ["function"]
          },
          "function": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "Name of the function to call"
              },
              "arguments": {
                "type": "string",
                "description": "JSON string of function arguments"
              }
            }
          }
        }
      },
      "ChatCompletionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the completion"
          },
          "object": {
            "type": "string",
            "enum": ["chat.completion"]
          },
          "created": {
            "type": "integer",
            "description": "Unix timestamp"
          },
          "model": {
            "type": "string",
            "description": "Model used"
          },
          "choices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatChoice"
            }
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        }
      },
      "ChatChoice": {
        "type": "object",
        "properties": {
          "index": {
            "type": "integer",
            "description": "Choice index"
          },
          "message": {
            "$ref": "#/components/schemas/ChatMessage"
          },
          "finish_reason": {
            "type": "string",
            "enum": ["stop", "length", "tool_calls", "content_filter"],
            "description": "Reason for completion"
          }
        }
      },
      "Usage": {
        "type": "object",
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "description": "Tokens in the prompt"
          },
          "completion_tokens": {
            "type": "integer",
            "description": "Tokens in the completion"
          },
          "total_tokens": {
            "type": "integer",
            "description": "Total tokens used"
          }
        }
      },
      "ChatCompletionStreamResponse": {
        "type": "object",
        "description": "Server-sent event stream response",
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string",
            "enum": ["chat.completion.chunk"]
          },
          "created": {
            "type": "integer"
          },
          "model": {
            "type": "string"
          },
          "choices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StreamChoice"
            }
          }
        }
      },
      "StreamChoice": {
        "type": "object",
        "properties": {
          "index": {
            "type": "integer"
          },
          "delta": {
            "$ref": "#/components/schemas/ChatMessage"
          },
          "finish_reason": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "Error message"
              },
              "type": {
                "type": "string",
                "description": "Error type"
              },
              "code": {
                "type": "string",
                "description": "Error code"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Optional chat API key for secured endpoints"
      }
    }
  }
}